Fix 32-bit build by static casting to size_t.

+++ test/complexity_test.cc
@@ -83,7 +83,7 @@ ADD_COMPLEXITY_CASES(big_o_1_test_name, rms_o_1_test_name, lambda_big_o_1);
 
 std::vector<int64_t> ConstructRandomVector(int64_t size) {
   std::vector<int64_t> v;
-  v.reserve(size);
+  v.reserve(static_cast<size_t>(size));
   for (int64_t i = 0; i < size; ++i) {
     v.push_back(std::rand() % size);
   }
